requirements.txt
numpy==1.24.4
opencv-python==4.6.0.66
packaging==23.1
Pillow==10.0.0
pypdf2==3.0.1
pytesseract==0.3.9
typing-extensions==4.7.1
